pyproject.toml
[project]
name = "nh-planner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Marcin Kostrzewa", email = "Manik24901@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "click>=8.1.8",
    "ollama>=0.4.5",
    "playwright>=1.49.1",
    "pydantic>=2.10.4",
    "python-dateutil>=2.9.0.post0",
    "rich>=13.9.4",
    "tqdm>=4.67.1",
]

[project.scripts]
nh-planner = "nh_planner.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "isort>=5.13.2",
    "ruff>=0.8.5",
]
